refactor: reorganize project into modules
